requirements.txt
aiohttp==3.8.5
aiohttp-client-cache==0.11.0
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.4.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
assemblyai==0.20.2
asttokens==2.4.1
async-timeout==4.0.3
attrbox==0.1.5
attrs==23.1.0
autogen==1.0.16
Automat==22.10.0
autopep8==2.1.0
av==10.0.0
backports.tarfile==1.1.1
beautifulsoup4==4.12.2
billiard==4.2.0
black==24.8.0
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.1
cattrs==23.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
closeio==2.1
coloredlogs==15.0.1
comm==0.2.1
constantly==15.1.0
contourpy==1.2.0
cryptography==43.0.1
cssselect==1.2.0
ctranslate2==3.24.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
docopt==0.6.2
docutils @ file:///private/tmp/docutils-20231211-37909-likvzd/docutils-0.20.1
elevenlabs==1.10.0
executing==2.0.1
faster-whisper==0.10.0
ffmpeg-python==0.2.0
filelock==3.12.3
FLAML==2.1.2
Flask==3.0.3
flatbuffers==23.5.26
flower==2.0.1
fonttools==4.48.1
fpdf==1.7.2
frozenlist==1.4.0
fsspec==2023.12.2
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gspread==5.12.4
gspread-dataframe==3.3.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.4
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib-metadata==7.0.1
incremental==22.10.0
install==1.3.5
ipykernel==6.29.2
ipython==8.21.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==25.2.1
kiwisolver==1.4.5
kombu==5.3.7
llvmlite==0.42.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
narwhals==1.9.4
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.17
numba==0.59.0
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.30.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parsel==1.8.1
parso==0.8.3
pathspec==0.12.1
pdfmerge==1.0.0
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.24.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyautogen==0.2.27
PyAutoGUI==0.9.54
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==8.2.2
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==3.9.0
pyperclip==1.8.2
pyppeteer==1.0.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
queuelib==1.6.2
readme_renderer==43.0
redis==5.0.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-cache==1.1.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
rubicon-objc==0.4.8
runstats==2.0.0
scipy==1.12.0
Scrapy==2.10.1
selenium==4.18.1
serpapi==0.1.5
service-identity==23.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.39.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tldextract==3.4.4
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.2
torch==2.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
twine==5.0.0
Twisted==22.10.0
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.18
vine==5.1.0
w3lib==2.1.2
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.0
zyte==0.0.0
zyte-api==0.4.8
